api/requirements.txt
# 后端API依赖
fastapi==0.104.1
uvicorn==0.24.0.post1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
